Break out helpers for readability
